# Reduce memory usage in imgutil package