--- a/go.mod
+++ b/go.mod
@@ -4,6 +4,5 @@ go 1.16
 
 require (
 	github.com/google/go-cmp v0.5.5
-	golang.org/x/image v0.0.0-20210220032944-ac19c3e999fb
 	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c
 )