… go.sum
… imgutil/testdata/wikipe-tan-100x123-downscale.png
… imgutil/testdata/wikipe-tan-100x123-upscale.png
… imgutil/testdata/wikipe-tan-100x123.png