As mentioned in #9, current image conversion pipeline uses a lot of memory by copying images instead of operating on them in-place. This PR aims to address this issue.