@naisuuuu 2021-04-10
Made some improvements:

old @ f22c060, new @ b03878d:
```
name             old time/op    new time/op    delta
AdjustGamma-12      106µs ± 1%     105µs ± 1%   -0.99%  (p=0.004 n=10+8)
AutoContrast-12     118µs ± 9%     126µs ± 5%   +6.69%  (p=0.007 n=10+10)

name             old alloc/op   new alloc/op   delta
AdjustGamma-12     49.5kB ± 0%     0.3kB ± 0%  -99.40%  (p=0.000 n=10+9)
AutoContrast-12    51.6kB ± 0%     2.4kB ± 0%  -95.31%  (p=0.000 n=9+8)

name             old allocs/op  new allocs/op  delta
AdjustGamma-12       5.00 ± 0%      3.00 ± 0%  -40.00%  (p=0.000 n=10+10)
AutoContrast-12      9.00 ± 0%      7.00 ± 0%  -22.22%  (p=0.000 n=10+10)
```
old @ f22c060, new @ 631984f:
```
name             old time/op    new time/op    delta
AdjustGamma-12      106µs ± 1%      73µs ± 8%   -31.37%  (p=0.000 n=10+10)
AutoContrast-12     118µs ± 9%      97µs ± 2%   -17.82%  (p=0.000 n=10+10)

name             old alloc/op   new alloc/op   delta
AdjustGamma-12     49.5kB ± 0%     0.0kB       -100.00%  (p=0.000 n=10+10)
AutoContrast-12    51.6kB ± 0%     2.1kB ± 0%   -95.88%  (p=0.000 n=9+8)

name             old allocs/op  new allocs/op  delta
AdjustGamma-12       5.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
AutoContrast-12      9.00 ± 0%      4.00 ± 0%   -55.56%  (p=0.000 n=10+10)
```

By looking at all the benchmarks @ f22c060 though, the biggest elephant in the room seems to be the inefficient `imgutil.Fit` implementation:
```
name                  alloc/op
Grayscale/RGBA-12     49.3kB ± 0%
Grayscale/RGBA64-12   49.3kB ± 0%
Grayscale/NRGBA-12    49.3kB ± 0%
Grayscale/NRGBA64-12  49.3kB ± 0%
Grayscale/YCbCr-12    49.2kB ± 0%
Grayscale/Gray-12     49.2kB ± 0%
AdjustGamma-12        49.5kB ± 0%
AutoContrast-12       51.6kB ± 0%
Fit-12                1.07MB ± 0% # this is a lot of memory

name                  allocs/op
Grayscale/RGBA-12       4.00 ± 0%
Grayscale/RGBA64-12     4.00 ± 0%
Grayscale/NRGBA-12      4.00 ± 0%
Grayscale/NRGBA64-12    4.00 ± 0%
Grayscale/YCbCr-12      2.00 ± 0%
Grayscale/Gray-12       2.00 ± 0%
AdjustGamma-12          5.00 ± 0%
AutoContrast-12         9.00 ± 0%
Fit-12                  12.0 ± 0%
```
I *could* also optimize some parts of `imgutil.Grayscale`, but that would only really make a difference for images which are already decoded as `image.Gray` (conversion between types necessitates copying at least parts of the underlying pixel array), and it would get in the way of stride correction. This may be worthwhile to do down the road anyway, as most images in a typical manga archive are grayscale and able to be decoded as such.