@naisuuuu 2021-04-13
Copied parts of `x/image/draw` into `imgutil` and modified them for usage with grayscale images specifically. This should be a temporary solution until https://github.com/golang/go/issues/44808 is implemented or grayscale dest fast paths are added to `x/image/draw`.

old @ f22c060, new @ c012e2b:
```
name    old time/op    new time/op    delta
Fit-12    1.26ms ± 1%    0.71ms ± 4%  -43.70%  (p=0.000 n=9+10)

name    old alloc/op   new alloc/op   delta
Fit-12    1.07MB ± 0%    0.30MB ± 0%  -72.23%  (p=0.000 n=10+10)

name    old allocs/op  new allocs/op  delta
Fit-12      12.0 ± 0%       8.0 ± 0%  -33.33%  (p=0.000 n=10+10)
```
Further optimization requires changing the `imgutil` api to take advantage of `Kernel.NewScaler`, WIP.