@naisuuuu 2021-04-14
old @ f22c060, new @ 6ce5af4:
```
name             old time/op    new time/op    delta
AdjustGamma-12      106µs ± 1%      73µs ± 4%   -31.17%  (p=0.000 n=10+9)
AutoContrast-12     118µs ± 9%      93µs ± 5%   -21.05%  (p=0.000 n=10+10)
Fit-12             1.26ms ± 1%    0.72ms ± 2%   -42.84%  (p=0.000 n=9+10)

name             old alloc/op   new alloc/op   delta
AdjustGamma-12     49.5kB ± 0%     0.0kB       -100.00%  (p=0.000 n=10+10)
AutoContrast-12    51.6kB ± 0%     2.1kB ± 0%   -95.88%  (p=0.000 n=9+8)
Fit-12             1.07MB ± 0%    0.30MB ± 0%   -72.23%  (p=0.000 n=10+10)

name             old allocs/op  new allocs/op  delta
AdjustGamma-12       5.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
AutoContrast-12      9.00 ± 0%      4.00 ± 0%   -55.56%  (p=0.000 n=10+10)
Fit-12               12.0 ± 0%       8.0 ± 0%   -33.33%  (p=0.000 n=10+10)
```
The `Grayscale` implementation remains mostly unchanged, due to issues explained in https://github.com/naisuuuu/mangaconv/pull/11#issuecomment-817136588.

This seems to be about as much as is possible without bigger API changes outside of the `imgutil` package.
Performance of `Fit` can be further improved by taking advantage of `*Kernel.NewScaler`, however this requires changes to how the image processing pipeline works in general. This should be worthwhile, as illustrated by benchmarks:
```
name             time/op
Fit-12            720µs ± 2%
KernelScaler-12   394µs ± 4%

name             alloc/op
Fit-12            296kB ± 0%
KernelScaler-12  18.8kB ± 1%

name             allocs/op
Fit-12             8.00 ± 0%
KernelScaler-12    2.00 ± 0%
```
This work will be continued in a following PR, this time optimizing the entire image processing pipeline instead of isolated `imgutil` functions.